# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bridge-Watch API",
    "description": "Real-time Stellar bridge monitoring API. Provides bridge statuses, asset health, liquidity, price feeds, alert management, analytics, and administrative controls.\n\n## Authentication\nProtected endpoints require an API key in the `x-api-key` header.\n\n## Rate Limiting\nAll endpoints are rate-limited per IP using a sliding-window algorithm backed by Redis. Exceeding the limit returns **429 Too Many Requests** with a `Retry-After` header.\n\n## Versioning\nAll REST endpoints are prefixed with `/api/v1/`. Breaking changes increment the version segment.\n\n## Error Format\n```json\n{ \"error\": \"Short label\", \"message\": \"Human-readable description\" }\n```",
    "version": "1.0.0",
    "contact": {
      "name": "Bridge-Watch Team",
      "url": "https://github.com/StellaBridge/Bridge-Watch"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" },
    { "url": "https://api.bridge-watch.io", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for protected endpoints"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/aggregation/{symbol}/prices": {
      "get": {
        "tags": ["Aggregation"],
        "summary": "Aggregate price data",
        "parameters": [
          { "name": "symbol", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "interval", "in": "query", "required": true, "schema": { "type": "string", "enum": ["1m", "5m", "15m", "1h", "4h", "1d"] } },
          { "name": "startTime", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } },
          { "name": "endTime", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } }
        ],
        "responses": { "200": { "description": "Price aggregations" } }
      }
    },
...
    "/api/v1/aggregation/{symbol}/volume": {
      "get": {
        "tags": ["Aggregation"],
        "summary": "Aggregate volume data",
        "parameters": [
          { "name": "symbol", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "interval", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "startTime", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } },
          { "name": "endTime", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } }
        ],
        "responses": { "200": { "description": "Volume aggregations" } }
      }
    },
...
  }
}